WIP: Add support for In-Band console diagnostics
Based on @vinzenz's earlier pull request #527.

When upgrade fails in initramfs, log files and other assets can be
tarballed and passed to ibdmp() function.  This function will encode
the tarball into Base64 and start spamming the console with pieces of
the Base64 plus some meta-data.

The encoding is designed to be resillient towards various forms of
corruption such as other things writing to the console (mid-line) or
line breaking (often seen on s390x machines in our Beaker pool).

This is achieved by lot of redundancy, so console log size will be
affected.  Following example transfers a very tiny tarball (leapp.db,
journalctl and some tiny files):

    6.4M console.log            # original; 3 dumps
    1.1M console.log.no-ibdmp   # with _ibdmp lines removed
    616K inband.tar.xz          # decoded tarball

Decoder (in Python 3) is included in utils/ subdirectory.

    utils/ibdmp-decode console.log ibdmp.tar.xz